feat: Add endpoint to serve user images by filename
This commit adds a new endpoint `serveImage` in the `UsersController` to serve user images based on the provided filename. The endpoint allows clients to retrieve and display user images by specifying the filename in the URL. This enhances the functionality of the application by providing a convenient way to access and display user images.