refactor: Update CreateUserDto to use string type for face descriptions
The CreateUserDto class in the src/users/dto/create-user.dto.ts file has been updated to use the string type for the faceDescription properties instead of the number type. This change ensures that the face descriptions are stored and processed as strings, which is the correct data type for this field.