feat: Add endpoint to serve assignment images
This commit adds a new endpoint `GET /assignments/image/filename/:filename` to the `AssignmentsController` that allows serving assignment images. The endpoint takes a `filename` parameter and sends the corresponding image file from the `assignment_images` directory as a response. This feature enhances the application by providing a way to retrieve and display assignment images.