feat: Update attendance image in update method
The `update` method in the `AttendancesController` has been updated to include the functionality to update the attendance image. If a file is uploaded, the `attendanceImage` attribute of the `updateAttendanceDto` is set to the uploaded file's filename. If no file is uploaded, the `attendanceImage` attribute is set to 'noimage.jpg'. This change enhances the update functionality by allowing users to update the attendance image along with other attendance details.